+++ mirage/projectstartup/angular_app_create.py
@@ -17,88 +17,110 @@ Copyright 2017-2018 Shota Shimazu.
 
 import os
 import shutil
-
 from mirage import proj
 from mirage import fileable
 from mirage.flow import Workflow
-from mirage.command import log, command
-from mirage.miragefile import source
-from mirage.template import readme_md
-from mirage.template import gitignore
-from mirage.template import package_json
+from mirage import system as mys
+from mirage.template import readme_md, gitignore, package_json
+from mirage.miragefile import source, source2
 
 
-class ReactStartupWorkFlow(Workflow):
-    
+class NgStartupWorkFlow(Workflow):
+
     def constructor(self):
-        self._project_name = None
+        self._plugins = self._option
 
 
     def main(self):
 
-        # Check 
+        # Check
         try:
             self._check_before()
         except:
             return
 
         # Input information
-        log("Please type your new Django application information.")
+        mys.log("Please type your new Django application information.")
 
         # Check namespace
         try:
-            self._project_name = log("Project name", withInput = True)
+            self._project_name = mys.log("Project name", withInput = True)
             self._check_namesapce(self._project_name)
         except:
-            log("Project \"{0}\" is already exists.".format(self._project_name), withError = True,
+            mys.log("Project \"{0}\" is already exists.".format(self._project_name), withError = True,
                     errorDetail = "Please remove duplication of Django project namespace.")
             return
 
-        version      = log("App version", withInput = True, default = "0.0.1")
-        author       = log("Author name", withInput = True)
-        email        = log("Email",       withInput = True)
-        git_url      = log("Git URL",     withInput = True)
-        license_name = log("License",     withInput = True)
-        description  = log("Description", withInput = True)
-        copyrightor  = log("Copyrightor", withInput = True, default = author)
+        version      = mys.log("App version", withInput = True, default = "0.0.1")
+        author       = mys.log("Author name", withInput = True)
+        email        = mys.log("Email",       withInput = True)
+        git_url      = mys.log("Git URL",     withInput = True)
+        license_name = mys.log("License",     withInput = True)
+        description  = mys.log("Description", withInput = True)
+        copyrightor  = mys.log("Copyrightor", withInput = True, default = author)
 
 
 
-        
+
         self._create_new_django_app()
 
+        # Create logging instance
+        logger = mys.Progress()
 
         with proj.InDir("./" + self._project_name):
-            
+
             # Generate .gitignore
-            log("Generating gitignore...")
+            #log("Generating gitignore...")
+            logger.write("Generating gitignore...", withLazy = True)
             self._create_template_git_project()
 
             # Generate README.md
-            log("Generating readme...")
-            self._create_docs()
+            logger.update("Generating readme...", withLazy = True)
+            self._create_docs(description)
 
             # Generate Miragefile
-            log("Generating Miragefile...")
+            logger.update("Generating Miragefile...", withLazy = True)
             self._create_miragefile(version, author, email, git_url, license_name, description, copyrightor)
 
             # Add remote repo
-            log("Adding remote repository...")
-            command("git remote add origin " + git_url)
+            logger.update("Adding remote repository...", withLazy = True)
+            mys.command("git remote add origin " + git_url)
 
-            # Create Angular App
-            log("Creating Angular app...")
+            # Create React App
             self._create_package_json()
-            command("yarn add --dev @angular/cli")
-            command("./node_modules/.bin/create-react-app --scripts-version=react-scripts-ts shell")
+            logger.update("Preparing Angular setup tools...", withLazy=True)
+            logger.update("Preparing Angular setup tools... ( please wait for a momenet )")
+            mys.command("yarn add --dev @angular/cli")
 
-            log("Installing additional packages...")
-            with proj.InDir("./shell"):
-                command("yarn add redux react-redux")
-                command("yarn add react-router react-router-dom")
+            # To avoid ng inside the project error, remove package.package.json
+            os.remove("package.json")
+
+            logger.update("Configuring Angular setup tools to use Yarn package...")
+            mys.command("./node_modules/.bin/ng set --global packageManager=yarn")
+            logger.update("Creating Angular app...", withLazy = True)
+            logger.update("Creating Angular app... ( please wait for a momenet )")
+            mys.command("./node_modules/.bin/ng new shell")
+
+
+            if self._plugins == "--nebular":
+                logger.update("Installing Nebular packages...", withLazy = True)
+                logger.update("Installing Nebular packages... ( please wait for a momenet )")
+                with proj.InDir("./shell"):
+                    mys.command("yarn add @nebular/theme @nebular/auth")
+
+            elif self._plugins == "--material":
+                logger.update("Installing material theme packages...")
+                with proj.InDir("./shell"):
+                    mys.command("yarn add @angular/material @angular/cdk")
+
+            elif self._plugins == "--bootstrap":
+                logger.update("Installing Bootstrap theme packages...")
+                with proj.InDir("./shell"):
+                    mys.command("yarn add ngx-bootstrap")
+
 
             # Cleaning
-            log("Cleaning...")
+            logger.update("Cleaning...", withLazy = True)
             fileable.rm("yarn.lock")
             fileable.rm("package.json")
             fileable.rm("node_modules/")
@@ -109,19 +131,22 @@ class ReactStartupWorkFlow(Workflow):
 
 
         # Completed
-        log("Completed!")
-    
+        logger.update("Completed!")
+
 
     def _create_new_django_app(self):
-        command("django-admin startproject " + self._project_name)
+        mys.command("django-admin startproject " + self._project_name)
 
 
-    def _create_miragefile(self, version, author, email, git_url, license_name, description, copyrightors):    
+    def _create_miragefile(self, version, author, email, git_url, license_name, description, copyrightors):
         with open("Miragefile", "w") as f:
             f.write(source.create(self._project_name, version, author, email, git_url, license_name, description, copyrightors))
-     
 
-    
+        with open("Miragefile@next.py", "w") as f:
+            f.write(source2.create(self._project_name, version, author, email, git_url, license_name, description, copyrightors))
+
+
+
     def _create_package_json(self):
         with open("package.json", "w") as f:
             f.write('{"name": "tmpapp", "version": "0.0.1"}')
@@ -133,20 +158,20 @@ class ReactStartupWorkFlow(Workflow):
         with open(".gitignore", "w") as f:
             f.write(ignorance)
 
-        command("git init")
+        mys.command("git init")
 
 
-    def _create_docs(self):
+    def _create_docs(self, description):
         with open("README.md", "a") as readme:
-            readme.write(readme_md.src(self._project_name))
+            readme.write(readme_md.src(self._project_name, description))
 
 
     def _check_before(self):
-        
+
         try:
             import django
         except ImportError:
-            log("Failed to import Django!", withError = True,
+            mys.log("Failed to import Django!", withError = True,
                                 errorDetail = "You have to install Django before creating a new Django project.")
             raise ImportError