+++ mirage/projectstartup/react_app_create.py
@@ -22,7 +22,7 @@ from mirage import fileable
 from mirage.flow import Workflow
 from mirage import system as mys
 from mirage.template import readme_md, gitignore, package_json
-from mirage.miragefile import source
+from mirage.miragefile import source, source2
 
 
 class ReactStartupWorkFlow(Workflow):
@@ -120,9 +120,12 @@ class ReactStartupWorkFlow(Workflow):
         mys.command("django-admin startproject " + self._project_name)
 
 
-    def _create_miragefile(self, version, author, email, git_url, license_name, description, copyrightors):    
+    def _create_miragefile(self, version, author, email, git_url, license_name, description, copyrightors):
         with open("Miragefile", "w") as f:
             f.write(source.create(self._project_name, version, author, email, git_url, license_name, description, copyrightors))
+
+        with open("Miragefile@next.py", "w") as f:
+            f.write(source2.create(self._project_name, version, author, email, git_url, license_name, description, copyrightors))